Organized files, split into multiple components
