--- components/eventDetails/EventDetails.tsx
@@ -1,311 +1,0 @@
-import { MaterialCommunityIcon } from "@components/common/Icons"
-import MenuDropdown from "@components/eventCard/MenuDropdown"
-import {
-  placemarkToAbbreviatedAddress,
-  placemarkToFormattedAddress
-} from "@lib/location"
-import { CurrentUserEvent, isAttendingEvent, isHostingEvent } from "@lib/events"
-import React from "react"
-import { Ionicons } from "@expo/vector-icons"
-import {
-  Image,
-  ImageBackground,
-  Pressable,
-  ScrollView,
-  StyleSheet,
-  Text,
-  View
-} from "react-native"
-import { BodyText, Caption, Headline, Title } from "@components/Text"
-import dayjs from "dayjs"
-import ExpandableText from "@components/common/ExpandableText"
-import { Divider, Icon } from "react-native-elements"
-import EventMapSnippet from "./EventMapSnippet"
-import { now } from "@lib/date"
-
-export type EventDetailsProps = {
-  event: CurrentUserEvent
-}
-
-const MARKER_SIZE = 44
-
-const EventDetails = ({ event }: EventDetailsProps) => {
-  const eventDuration = event.dateRange
-  const startDateFormat = eventDuration.formattedDate(
-    now(),
-    dayjs(eventDuration.startDate)
-  )
-  const endDateFormat = eventDuration.formattedDate(
-    now(),
-    dayjs(eventDuration.endDate)
-  )
-  const startTimeFormat = dayjs(eventDuration.startDate).format("h A")
-  const endTimeFormat = dayjs(eventDuration.endDate).format("h A")
-  const isFriend = false
-
-  return (
-    <ScrollView
-      style={[styles.container, styles.spacing]}
-      // scrollEnabled={true}
-      contentContainerStyle={{ flex: 1 }}
-    >
-      <Title>{event.title}</Title>
-      <View style={[styles.flexRow, { marginVertical: 16 }]}>
-        <Image
-          style={[styles.image]}
-          source={require("../../assets/icon.png")}
-        />
-        <View style={{ flex: 1 }}>
-          <View style={styles.flexRow}>
-            <Headline>{event.host.username}</Headline>
-            {!isFriend
-              ? (
-                <View style={[styles.flexRow, { alignItems: "center" }]}>
-                  <Ionicons
-                    style={{ marginHorizontal: 4 }}
-                    size={6}
-                    name="ellipse"
-                  />
-                  <Headline style={{ color: event.color }}>Add Friend</Headline>
-                </View>
-              )
-              : null}
-          </View>
-          <Caption>{event.host.handle}</Caption>
-        </View>
-      </View>
-
-      <View style={styles.iconSection}>
-        <View style={[styles.flexRow, styles.paddingIconSection]}>
-          <View style={{ justifyContent: "center" }}>
-            <Ionicons
-              style={[styles.iconStyling, { backgroundColor: event.color }]}
-              name="calendar"
-              color={"white"}
-              size={24}
-            />
-          </View>
-          <View style={styles.spacing}>
-            {eventDuration.endSameDay()
-              ? (
-                <View style={{ marginBottom: 4 }}>
-                  <Headline>{`${startDateFormat}`}</Headline>
-                  <Caption>{`from ${startTimeFormat} - ${endTimeFormat}`}</Caption>
-                </View>
-              )
-              : (
-                <View style={{ marginBottom: 4 }}>
-                  <Headline>{`From ${startDateFormat}, ${startTimeFormat}`}</Headline>
-                  <Caption>{`to ${endDateFormat}, ${endTimeFormat}`}</Caption>
-                </View>
-              )}
-            <Caption style={[{ color: event.color }, styles.captionLinks]}>
-              Add to Calendar
-            </Caption>
-          </View>
-        </View>
-
-        <Divider style={styles.divider} />
-
-        <View style={[styles.flexRow, styles.paddingIconSection]}>
-          <View style={{ justifyContent: "center" }}>
-            <Ionicons
-              style={[styles.iconStyling, { backgroundColor: event.color }]}
-              name="location"
-              color={"white"}
-              size={24}
-            />
-          </View>
-          {event.placemark
-            ? (
-              <View style={styles.spacing}>
-                <View style={{ marginBottom: 4 }}>
-                  <Headline>{event.placemark.name}</Headline>
-                  <View style={styles.flexRow}>
-                    {event.placemark.streetNumber && (
-                      <Caption>{`${event.placemark.streetNumber} `}</Caption>
-                    )}
-                    {event.placemark.street && (
-                      <Caption>{`${event.placemark.street}, `}</Caption>
-                    )}
-                    {event.placemark.city && (
-                      <Caption>{`${event.placemark.city}, `}</Caption>
-                    )}
-                    {event.placemark.region && (
-                      <Caption>{`${event.placemark.region}`}</Caption>
-                    )}
-                  </View>
-                </View>
-                <View style={styles.flexRow}>
-                  <Caption
-                    style={[
-                      { color: event.color, marginRight: 16 },
-                      styles.captionLinks
-                    ]}
-                  >
-                  Copy Address
-                  </Caption>
-                  <Caption style={[{ color: event.color }, styles.captionLinks]}>
-                  Directions
-                  </Caption>
-                </View>
-              </View>
-            )
-            : (
-              <View style={styles.spacing}>
-                <View style={{ marginBottom: 4 }}>
-                  <Headline>{`${event.coordinates.latitude}, ${event.coordinates.longitude}`}</Headline>
-                  <View style={styles.flexRow}>
-                    <Caption>Unknown Address</Caption>
-                  </View>
-                </View>
-                <View style={styles.flexRow}>
-                  <Caption
-                    style={[
-                      { color: event.color, marginRight: 16 },
-                      styles.captionLinks
-                    ]}
-                  >
-                  Copy Coordinates
-                  </Caption>
-                  <Caption style={[{ color: event.color }, styles.captionLinks]}>
-                  Directions
-                  </Caption>
-                </View>
-              </View>
-            )}
-        </View>
-
-        <Divider style={styles.divider} />
-
-        <View style={[styles.flexRow, styles.paddingIconSection]}>
-          <View style={{ justifyContent: "center" }}>
-            <Ionicons
-              style={[styles.iconStyling, { backgroundColor: event.color }]}
-              name="people"
-              color={"white"}
-              size={24}
-            />
-          </View>
-          <View style={styles.spacing}>
-            <Headline>{`${event.attendeeCount} Attending`}</Headline>
-            <Caption>View all attendees</Caption>
-          </View>
-          <View
-            style={[styles.flexRow, { flex: 1, justifyContent: "flex-end" }]}
-          >
-            <Ionicons
-              name="chevron-forward"
-              size={20}
-              style={{ alignSelf: "center", opacity: 0.3 }}
-            />
-          </View>
-        </View>
-
-        <Divider style={styles.divider} />
-
-        <View style={[styles.flexRow, styles.paddingIconSection]}>
-          <View style={{ justifyContent: "center" }}>
-            <Ionicons
-              style={[styles.iconStyling, { backgroundColor: event.color }]}
-              name="chatbox-ellipses"
-              color={"white"}
-              size={24}
-            />
-          </View>
-          <View style={styles.spacing}>
-            <Headline>Event Chat</Headline>
-            {isAttendingEvent(event.userAttendeeStatus)
-              ? (
-                <Caption>View the chat</Caption>
-              )
-              : (
-                <Caption>Join this event to access the chat</Caption>
-              )}
-          </View>
-        </View>
-      </View>
-
-      {event.description && (
-        <View style={{ marginTop: 16 }}>
-          <Headline style={{ marginBottom: 4 }}>About</Headline>
-          <ExpandableText
-            props={{ style: { color: event.color, marginTop: 5 } }}
-            text={event.description}
-            linesToDisplay={3}
-          />
-        </View>
-      )}
-
-      <View style={{ marginTop: 16 }}>
-        <Headline style={{ marginBottom: 8 }}>Location</Headline>
-        <EventMapSnippet
-          style={{ width: "100%", height: "50%" }}
-          minZoomLevel={12}
-          initialRegion={{
-            latitude: event.coordinates.latitude,
-            longitude: event.coordinates.longitude,
-            latitudeDelta: 0.1,
-            longitudeDelta: 0.1
-          }}
-          renderMarker={() => (
-            <Image
-              source={require("../../assets/icon.png")}
-              style={{
-                width: MARKER_SIZE,
-                height: MARKER_SIZE,
-                borderWidth: 1,
-                borderColor: "white",
-                borderRadius: 20
-              }}
-            />
-          )}
-          marker={{ key: event.id, location: event.coordinates }}
-        />
-      </View>
-    </ScrollView>
-  )
-}
-
-export default EventDetails
-
-const styles = StyleSheet.create({
-  container: {
-    paddingVertical: 24,
-    backgroundColor: "white"
-  },
-  flexRow: {
-    flexDirection: "row"
-  },
-  spacing: {
-    paddingHorizontal: 16
-  },
-  image: {
-    width: 40,
-    height: 40,
-    marginRight: 16,
-    borderRadius: 20
-  },
-  iconSection: {
-    backgroundColor: "#0000000D",
-    borderRadius: 8,
-    paddingVertical: 16
-  },
-  captionLinks: {
-    opacity: 1,
-    fontWeight: "bold"
-  },
-  divider: {
-    marginVertical: 16,
-    width: "80%",
-    height: 1,
-    alignSelf: "flex-end"
-  },
-  paddingIconSection: {
-    paddingHorizontal: 16
-  },
-  iconStyling: {
-    padding: 6,
-    borderRadius: 12
-  }
-})
