+++ screens/EventDetails/EventMapSnippet.tsx
